requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
redis==3.5.3
celery==3.1.18
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
uuid==1.30
